Suppress `yes` broken pipe output when generating SSH keys
On Linux in CI, the following error appears when running this post-up hook:

yes: standard output: Broken pipe

Normally, `yes` would be killed by the `SIGPIPE` signal that is
generated when it tries to write to the pipe when `ssh-keygen` has
exited. However, the CI traps this signal and instead `yes` detects this
error, thus reporting it. Sending stderr to /dev/null should be
sufficient in hiding this error.